move json schema generation for tool use into transformer-core
