+++ packages/amplify-graphql-transformer-core/src/utils/ai/graphql-scalar-json-schema-definitions.ts
@@ -1,10 +1,10 @@
-import { JSONSchema } from './graphql-json-schema-type';
+import { JSONSchema } from './json-schema';
 
 /**
  * JSON Schema definitions for GraphQL scalar types and AWS AppSync custom scalar types.
  * These definitions are used to create valid JSON schema for tool definitions in the
- * context of AI model interactions for generation routes, ensuring that generated responses
- * conform to the expected GraphQL types.
+ * context of AI model interactions for generation and converssation routes, ensuring that generated responses
+ * conform to the expected GraphQL types or inputs.
  *
  * Each constant represents a JSON Schema object that describes the structure and
  * constraints of a specific GraphQL scalar type. These scalar types are utilized when
