+++ packages/amplify-graphql-transformer-core/src/utils/ai/index.ts
@@ -1,0 +1,7 @@
+export type { JSONSchema } from './json-schema';
+export {
+  isDisallowedScalarType,
+  supportedScalarTypes,
+  GraphQLScalarJSONSchemaDefinition,
+} from './graphql-scalar-json-schema-definitions';
+export { convertNamedTypeToJSONSchema } from './named-type-conversion';
