+++ packages/amplify-graphql-transformer-core/src/utils/ai/json-schema.ts
@@ -1,0 +1,18 @@
+export type JSONSchema = {
+  type: string;
+  properties?: Record<string, JSONSchema>;
+  required?: string[];
+  items?: JSONSchema;
+  enum?: (string | number | boolean | null)[];
+  minimum?: number;
+  maximum?: number;
+  minLength?: number;
+  maxLength?: number;
+  pattern?: string;
+  format?: string;
+  description?: string;
+  default?: JSONLike;
+  additionalProperties?: boolean | JSONSchema;
+};
+
+type JSONLike = string | number | boolean | null | { [key: string]: JSONLike } | JSONLike[];
